Move created C protobuf code from 'core' to 'core/pb'
