+++ core/modules/dump.h
@@ -32,7 +32,7 @@
 #define BESS_MODULES_DUMP_H_
 
 #include "../module.h"
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 
 class Dump final : public Module {
  public:
