+++ core/modules/ether_encap.h
@@ -32,7 +32,7 @@
 #define BESS_MODULES_ETHERENCAP_H_
 
 #include "../module.h"
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 
 class EtherEncap final : public Module {
  public:
