+++ core/modules/generic_encap.h
@@ -32,7 +32,7 @@
 #define BESS_MODULES_GENERICENCAP_H_
 
 #include "../module.h"
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 
 #define MAX_FIELDS 8
 #define MAX_FIELD_SIZE 8
