+++ core/modules/hash_lb.h
@@ -32,7 +32,7 @@
 #define BESS_MODULES_HASHLB_H_
 
 #include "../module.h"
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 
 #define MAX_HLB_GATES 16384
