+++ core/modules/ip_encap.h
@@ -32,7 +32,7 @@
 #define BESS_MODULES_IPENCAP_H_
 
 #include "../module.h"
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 
 class IPEncap final : public Module {
  public:
