+++ core/modules/measure.h
@@ -32,7 +32,7 @@
 #define BESS_MODULES_MEASURE_H_
 
 #include "../module.h"
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 #include "../utils/histogram.h"
 #include "../utils/random.h"
