+++ core/modules/queue_inc.h
@@ -32,7 +32,7 @@
 #define BESS_MODULES_QUEUEINC_H_
 
 #include "../module.h"
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 #include "../port.h"
 
 class QueueInc final : public Module {
