+++ core/modules/queue_out.h
@@ -32,7 +32,7 @@
 #define BESS_MODULES_QUEUEOUT_H_
 
 #include "../module.h"
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 #include "../port.h"
 
 class QueueOut final : public Module {
