+++ core/modules/replicate.h
@@ -31,7 +31,7 @@
 #define BESS_MODULES_REPLICATE_H_
 
 #include "../module.h"
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 
 class Replicate final : public Module {
  public:
