+++ core/modules/rewrite.h
@@ -32,7 +32,7 @@
 #define BESS_MODULES_REWRITE_H_
 
 #include "../module.h"
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 
 class Rewrite final : public Module {
  public:
