+++ core/modules/split.h
@@ -32,7 +32,7 @@
 #define BESS_MODULES_SPLIT_H_
 
 #include "../module.h"
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 
 class Split final : public Module {
  public:
