+++ core/modules/timestamp.h
@@ -32,7 +32,7 @@
 #define BESS_MODULES_TIMESTAMP_H_
 
 #include "../module.h"
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 
 class Timestamp final : public Module {
  public:
