+++ core/modules/vxlan_decap.h
@@ -32,7 +32,7 @@
 #define BESS_MODULES_VXLANDECAP_H_
 
 #include "../module.h"
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 
 class VXLANDecap final : public Module {
  public:
