+++ core/modules/wildcard_match.h
@@ -36,7 +36,7 @@
 #include <rte_config.h>
 #include <rte_hash_crc.h>
 
-#include "../module_msg.pb.h"
+#include "../pb/module_msg.pb.h"
 #include "../utils/cuckoo_map.h"
 
 using bess::utils::HashResult;
